requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.1.0
Pillow>=10.0.0
